% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_spec.R
\name{as_rclaimlab_workflow}
\alias{as_rclaimlab_workflow}
\title{Convert a supported object to a workflow}
\usage{
as_rclaimlab_workflow(x, ...)
}
\arguments{
\item{x}{Object to convert.}

\item{...}{Conversion arguments.}
}
\description{
Convert a supported object to a workflow
}
