% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_receipt.R
\name{read_workflow_receipt}
\alias{read_workflow_receipt}
\title{Read a workflow receipt}
\usage{
read_workflow_receipt(path = ".")
}
\arguments{
\item{path}{Receipt JSON path or compiled workflow directory.}
}
\description{
Read a workflow receipt
}
