% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_engine.R
\name{run_workflow}
\alias{run_workflow}
\title{Execute an approved workflow locally in R}
\usage{
run_workflow(workflow)
}
\arguments{
\item{workflow}{Approved workflow specification.}
}
\value{
An `rclaimlab_workflow_run` containing an evidence bundle.
}
\description{
Execute an approved workflow locally in R
}
