% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_receipt.R
\name{validate_workflow_receipt}
\alias{validate_workflow_receipt}
\title{Validate a workflow receipt}
\usage{
validate_workflow_receipt(receipt)
}
\arguments{
\item{receipt}{Receipt object or JSON path.}
}
\description{
Validate a workflow receipt
}
