% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesson_manifest.R
\name{read_learning_receipt}
\alias{read_learning_receipt}
\title{Read an R-ClaimLab learning receipt}
\usage{
read_learning_receipt(path = ".")
}
\arguments{
\item{path}{Path to a receipt JSON file or a lesson directory.}
}
\value{
An object of class `rclaimlab_receipt`.
}
\description{
Read an R-ClaimLab learning receipt
}
